requirements.txt
browserstack_sdk==1.28.0
googletrans==4.0.0rc1
pytest==8.4.1
python-dotenv==1.1.0
requests==2.32.4
webdriver-manager==4.0.2
PyYAML==6.0.2
